Make cloud integrations type optional (#484)
## Provider

### New
- No new features or additions.

### Fixed
- No resolved bugs or issues.

### Changed
- The `type` field for the Integration is now optional instead of
required. It has a default value of "aws" and is marked as deprecated,
indicating it will be removed in a future version.

## Additional Information
The change to make the `type` field optional allows for greater
flexibility in integration configurations while maintaining backward
compatibility for existing users. Users should be aware that this field
is deprecated and will be removed in future releases. Please update your
configurations accordingly.