Add support for CF template 1.2.0 feature flags (#485)
## Provider

### New
- Added support for the following feature flags in AWS Cloud
Integrations:
- `allow_s3_access`: Boolean to allow Cloud Integration to access S3
buckets for Log Integrations.
- `enable_eks_autodiscovery`: Boolean to enable resource autodiscovery
for EKS clusters.
- `enable_rds_autodiscovery`: Boolean to enable resource autodiscovery
for RDS instances.
- `enable_redshift_autodiscovery`: Boolean to enable resource
autodiscovery for Redshift clusters.
- Added the following read-only / computed fields to the
`formal_integration_cloud` resource, so that it's easier to pass to the
CloudFormation stack resource:
- `aws_allow_s3_access` (Boolean) Whether AWS S3 access is allowed or
not.
- `aws_enable_eks_autodiscovery` (Boolean) Whether AWS EKS autodiscovery
is enabled or not.
- `aws_enable_rds_autodiscovery` (Boolean) Whether AWS RDS autodiscovery
is enabled or not.
- `aws_enable_redshift_autodiscovery` (Boolean) Whether AWS Redshift
autodiscovery is enabled or not.

### Fixed
- No fixed issues in this PR.

### Changed
- Updated the CloudFormation integration example to use template version
1.2.0.
- Adjusted resource definitions to include new feature flags and their
corresponding descriptions in schema.

## Additional Information
These changes enhance the flexibility and capabilities of our Cloud
Integration, allowing users to better manage AWS resources through
improved autodiscovery and access controls. Please review the
documentation for detailed usage instructions on the new parameters.