TEAMENG-221: path-scoped discovery schedules (#502)
## Provider

### New
- Introduced a new optional parameter `path` for the inventory object in
the Data Discovery resource.

### Fixed
- No bugs or issues were resolved in this PR.

### Changed
- Updated the Data Discovery resource to include the `path` parameter in
both creation and update functions.
- Adjusted the documentation to reflect the addition of the `path`
parameter.

## Additional Information
This change enhances the Data Discovery resource by allowing users to
specify a path for the inventory object, thereby improving the
flexibility and usability of the resource. The updates are reflected in
both the codebase and the documentation to ensure clarity for future
implementations.